refactor: 패키지명 수정 및 이동
- blackjackgame 패키지 -> game로 이름 수정
- dealer, player 패키지 -> participant 패키지로 이동
